requirements.txt
Django==1.9.2
djangorestframework==3.3.2
markdown==2.6.5
django-crispy-forms==1.6.0
django-filter==0.12.0
django-guardian==1.4.1